Ad-hoc secret management for the bearer token
Abstracts the secret away from the more changeable index.js, where all
the client-side training tasks are implemented. Inspired by this post:
https://www.simonewebdesign.it/how-to-get-environment-variables-in-the-browser/

This solution works best when marking these env.js files with
`$ git update-index --skip-worktree FILENAME`
so that git doesn't track changes on them. More on this here:
https://compiledsuccessfully.dev/git-skip-worktree/
But I can't do this on my fork because switching branches gets messy:
https://stackoverflow.com/q/29549318/11363646

I added `**/env.js` to the .gitignore as a safeguard for anyone that
tinkers with the directory structure (like me), so we won't unwittingly
push our bearer tokens to a public repo.